\docType{data}
\name{CD8}
\alias{CD8}
\title{Set of 10,000 beta-chain CDR3s derived from the T cell receptors of
human naive CD8+ T cells}

\format{
\code{data.frame} with 10,000 rows and 1 column 'CDR3b'.
}
\source{
http://50.255.35.37:8080/
}
\usage{
data(CD8)
}
\description{
Contains 10,000 sequences of beta-chain complementarity determining regions 3
(CDR3b). This data set was drawn from a much larger set of beta-chain CDR3s
derived from the T cell receptors of human naive CD8+ T cells
(source: http://50.255.35.37:8080)
}

\examples{
data("CD8")
}

\keyword{datasets}
