% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{check_data_ref}
\alias{check_data_ref}
\title{Input checks for reference database}
\usage{
check_data_ref(data_ref)
}
\arguments{
\item{data_ref, }{data.frame: reference database}
}
\description{
Input checks for reference database
}
