% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{check_data_sample}
\alias{check_data_sample}
\title{Input checks for TCR sample}
\usage{
check_data_sample(data_sample)
}
\arguments{
\item{data_sample, }{data.frame: TCR sample}
}
\description{
Input checks for TCR sample
}
