% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{check_flank_size}
\alias{check_flank_size}
\title{Input checks for flank size}
\usage{
check_flank_size(flank_size)
}
\arguments{
\item{flank_size, }{flank size}
}
\description{
Input checks for flank size
}
