% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{check_global_max_dist}
\alias{check_global_max_dist}
\title{Input checks for maximal global distance}
\usage{
check_global_max_dist(global_max_dist)
}
\arguments{
\item{global_max_dist, }{maximal global distance}
}
\description{
Input checks for maximal global distance
}
