% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{check_global_pairs}
\alias{check_global_pairs}
\title{Input checks for global pairs and sample data}
\usage{
check_global_pairs(global_pairs, data_sample)
}
\arguments{
\item{global_pairs, }{global pairs}

\item{data_sample, }{sample data}
}
\description{
Input checks for global pairs and sample data
}
