% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{check_ks_local_min_ove}
\alias{check_ks_local_min_ove}
\title{Input checks for motif lengths and local minimum ove}
\usage{
check_ks_local_min_ove(ks, local_min_ove)
}
\arguments{
\item{ks, }{vector of integers: motif lengths to use (default ks=(2,3,4))}

\item{local_min_ove, }{local minimum ove}
}
\description{
Input checks for motif lengths and local minimum ove
}
