% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{check_local_min_o}
\alias{check_local_min_o}
\title{Input checks for minimal local o}
\usage{
check_local_min_o(local_min_o)
}
\arguments{
\item{local_min_o, }{minimal local o}
}
\description{
Input checks for minimal local o
}
