% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{check_local_min_p}
\alias{check_local_min_p}
\title{Input checks for minimal local p-value}
\usage{
check_local_min_p(local_min_p)
}
\arguments{
\item{local_min_p, }{minimal local p-value}
}
\description{
Input checks for minimal local p-value
}
