% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{check_trim_flanks}
\alias{check_trim_flanks}
\title{Input checks for trim flanks}
\usage{
check_trim_flanks(trim_flanks)
}
\arguments{
\item{trim_flanks, }{trim flanks}
}
\description{
Input checks for trim flanks
}
