% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{check_trim_flanks_flank_size}
\alias{check_trim_flanks_flank_size}
\title{Input checks for trim flanks and flank size}
\usage{
check_trim_flanks_flank_size(trim_flanks, flank_size)
}
\arguments{
\item{trim_flanks, }{trim flanks}

\item{flank_size, }{flank size}
}
\description{
Input checks for trim flanks and flank size
}
