% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{check_version}
\alias{check_version}
\title{Input checks for gliph version}
\usage{
check_version(version)
}
\arguments{
\item{version, }{integer: version = 1, 2 or 3, gliph version to use}
}
\description{
Input checks for gliph version
}
