\name{get_beta_scatterplot}
\alias{get_beta_scatterplot}
\title{Compare community \eqn{\beta}s between pairs of repertoires}
\description{
Visualize the \eqn{\beta} means as a 2D scatterplot, representing relative 
community occupancies for all pairs of repertoires. At the same time, annotate 
the communities (dots) based on their specificity.
}
\details{
The user has to provide a vector of antigen species (e.g. 
\code{ag_species} = c("EBV", "CMV")) and/or a vector of antigen 
genes (e.g. \code{ag_genes} = "MLANA"). Furthermore, the user has 
to provide nodes (\code{node_summary} data.frame created by the 
function \code{detect_communities}) and \code{beta} data.frame 
which is part of \code{posterior_summary} generated by the function 
\code{dco}. 

The user can also select an annotation database \code{db},
such as "vdjdb", "mcpas" or "tcr3d"; and restrict the annotation to
specific IR chains, such as "CDR3a", "CDR3b" or "both". By default,
we will look for perfect matches (\code{db_dist=0}) between CDR3 
sequences in the input and in the annotation database for annotation. 
Flexible annotation based on edit distances can be performed by 
increasing \code{db_dist}.
}
\usage{
get_beta_scatterplot(beta,
                     node_summary,
                     ag_species, 
                     ag_genes,
                     db = "vdjdb",
                     db_dist = 0,
                     chain = "both")
}
\arguments{
\item{beta}{\code{beta} data.frame}
\item{node_summary}{\code{node_summary} data.frame}
\item{ag_species}{antigen species, character vector, e.g. c("EBV", "CMV")}
\item{ag_genes}{antigen genes, character vector, e.g. "MLANA"}
\item{db}{annotation database, character, e.g. "vdjdb"}
\item{db_dist}{maximum edit distance threshold for matching, nummeric}
\item{chain}{immune receptor chain for annotation, "both", "CDR3a" or "CDR3b"}
}
\value{
The output is a list with 4 elements: 

\code{node_annotations}: annotated \code{node_summary} 

\code{beta_summary}: annotated \code{beta} 

\code{vars}: annotation variables

\code{scatterplots}: a list of scatterplots: Each element of the list contains 
a scatterplots for a specific antigen species/gene. Within each element of the 
list there are \eqn{n^2} panels (comparisons between repertoire pairs), with 
\eqn{n} as the number of repertoires.
}
\examples{
# load package input data
data("CDR3ab", package = "ClustIRR")
a <- data.frame(CDR3a = CDR3ab[1:500, "CDR3a"],
                  CDR3b = CDR3ab[1:500, "CDR3b"],
                  clone_size = 1,
                  sample = "a")

b <- data.frame(CDR3a = CDR3ab[401:900, "CDR3a"],
                  CDR3b = CDR3ab[401:900, "CDR3b"],
                  clone_size = 1,
                  sample = "b")
b$clone_size[1] <- 20

# run ClustIRR analysis
c <- clustirr(s = rbind(a, b))

# detect communities
gcd <- detect_communities(graph = c$graph, 
                          algorithm = "leiden",
                          resolution = 1,
                          weight = "ncweight",
                          chains = c("CDR3a", "CDR3b"))

# differential community occupancy analysis
dco <- dco(community_occupancy_matrix = gcd$community_occupancy_matrix,
           mcmc_control = list(mcmc_chains = 2, 
                               mcmc_cores = 2))

# generate beta violin plots
beta_scatterplot <- get_beta_scatterplot(beta = dco$posterior_summary$beta,
                                         node_summary = gcd$node_summary,
                                         ag_species = c("EBV", "CMV"),
                                         ag_genes = "MLANA",
                                         db = "vdjdb",
                                         db_dist = 0,
                                         chain = "both")
}
