\name{get_graph}
\alias{get_graph}
\title{Get graph structure from \code{clust_irr} object}
\usage{get_graph(clust_irr, sample_id = "S", custom_db = NULL)}
\arguments{
\item{clust_irr}{S4 object generated by the function \code{cluster_irr}}
\item{sample_id}{character, name of the repertoire (default = S)}
\item{custom_db}{data.frame, custom database with CDR3 sequences and their 
matching antigens. The structure of custom_db must be identical to data(vdjdb).
This is an optional input to annotate the clones. If \code{custom_db} is not 
provided, ClustIRR will use internal databases. See description below.}
}
\value{
The main output of this function is an \code{igraph} object. Additionally, 
we provide a data.frame in which rows are clones (vertices) in the graph.
}

\description{
The main output of this function is an \code{igraph} object.

The vertices in the graph represent clones. Undirected edges are drawn 
between a pair of vertices if the corresponding clones that are locally 
and/or globally similar.

The function matches CDR3s from the clones in the graph with CDR3 with known 
epitopes from these databases: VDJdb, McPAS-TCR and TCR3d. With the optional 
input parameter \code{custom_db}, the user can provide additional databases.
}

\examples{
# load package input data
data("CDR3ab")
s <- data.frame(CDR3b = CDR3ab[1:100, "CDR3b"], clone_size = 1)
r <- data.frame(CDR3b = CDR3ab[1:5000, "CDR3b"], clone_size = 1)

# artificially enrich motif 'RWGW' inside sample dataset
substr(x = s$CDR3b[1:20], start = 6, stop = 9) <- "RWGW"

# add an artificial clonal expansion of two sequences to the sample dataset
s <- rbind(s, data.frame(CDR3b = c("CATSRADKPDGLDALETQYF", 
                                   "CATSRAAKPDGLAALSTQYF"),
                         clone_size = 5))

# run ClustIRR analysis
out <- cluster_irr(s = s,
                   r = r,
                   ks = 4,
                   cores = 1,
                   control = list(trim_flank_aa = 3))

# get graph
g <- get_graph(out)

names(g)
}