% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gliph.R
\name{gliph}
\alias{gliph}
\title{Find groups of TCRs with locally or globally similar CDR3s}
\usage{
gliph(
  data_sample,
  data_ref,
  version = 2,
  ks = c(2, 3, 4),
  cores = 1,
  control = list(B = 1000, global_max_dist = 1, local_max_fdr = 0.05, local_min_ove = 2,
    local_min_o = 3, trim_flank_aa = 0, global_pairs = NULL, low_mem = FALSE)
)
}
\arguments{
\item{data_sample, }{data.frame: TCR sample}

\item{data_ref, }{data.frame: reference database}

\item{version, }{integer: version = 1, 2 or 3, gliph version to use}

\item{ks, }{vector of integers: motif lengths to use (default ks=(2,3,4)),}

\item{cores, }{integer: number of CPU cores to use}

\item{control, }{list: auxiliarry input parameters (described below)}
}
\value{
list(clust = clust, edges = edges, data_sample = data_sample)
}
\description{
Find groups of TCRs with locally or globally similar CDR3s
}
\examples{
## TODO add meaningful examples
print("example")
}
