% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs_CD4_ref}
\alias{hs_CD4_ref}
\title{HS CD4 ref}
\format{
An object of class \code{data.frame} with 772312 rows and 6 columns.
}
\usage{
data(hs_CD4_ref)
}
\description{
\code{"hs_CD4_ref"} to be described
}
\keyword{datasets}
