% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs_CD8_ref}
\alias{hs_CD8_ref}
\title{HS CD8 ref}
\format{
An object of class \code{data.frame} with 573211 rows and 6 columns.
}
\usage{
data(hs_CD8_ref)
}
\description{
\code{"hs_CD8_ref"} to be described
}
\keyword{dataset}
