% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm_CD4_ref}
\alias{mm_CD4_ref}
\title{MM CD4 ref}
\format{
An object of class \code{data.frame} with 22268 rows and 6 columns.
}
\usage{
data(mm_CD4_ref)
}
\description{
\code{"mm_CD4_ref"} to be described
}
\keyword{datasets}
