% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm_CD8_ref}
\alias{mm_CD8_ref}
\title{MM CD8 ref}
\format{
An object of class \code{data.frame} with 55847 rows and 6 columns.
}
\usage{
data(mm_CD8_ref)
}
\description{
\code{"mm_CD8_ref"} to be described
}
\keyword{datasets}
