% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{parameter_check}
\alias{parameter_check}
\title{Input checks for gliphR parameters}
\usage{
parameter_check(data_sample, data_ref, version, ks, cores, control)
}
\arguments{
\item{data_sample, }{data.frame: TCR sample}

\item{data_ref, }{data.frame: reference database}

\item{version, }{integer: version = 1, 2 or 3, gliph version to use}

\item{ks, }{vector of integers: motif lengths to use (default ks=(2,3,4)),}

\item{cores, }{integer: number of CPU cores to use}

\item{control, }{list: auxiliary input parameters}
}
\description{
Input checks for gliphR parameters
}
