\docType{data}
\name{CD8}
\alias{CD8}
\title{Set of 10,000 beta-chain CDR3s derived from the T cell receptors of
human naive CD8+ T cells}

\format{
\code{data.frame} with 10,000 rows and 1 column \code{CDR3b}.
}
\source{
\href{http://50.255.35.37:8080/}{GLIPH version 2}
}
\usage{
data(CD8)
}
\description{
Contains 10,000 sequences of beta-chain complementarity determining regions 3
(\emph{CDR3b}). This data set was drawn from a much larger set of 
\eqn{\beta}-chain CDR3s derived from the T cell receptors of human naive CD8+ 
T cells (source: \href{http://50.255.35.37:8080/}{GLIPH version 2})
}

\examples{
data("CD8")
}

\keyword{datasets}
