\docType{data}
\name{CDR3ab}
\alias{CDR3ab}
\title{Mock data set of complementarity determining region 3 (CDR3)
sequences from the \eqn{\alpha} and \eqn{\beta} chains of 10,000 T cell
receptors}

\format{
\code{data.frame} with 10,000 rows and 2 columns \code{CDR3a} and \code{CDR3b}.
}
\source{
\href{http://50.255.35.37:8080/}{GLIPH version 2}
}
\usage{
data(CDR3ab)
}
\description{
Mock data set containing amino acid sequences of paired CDR3s from the
\eqn{\alpha} and \eqn{\beta} chains of 10,000 T cell receptors. All CDR3
sequences were drawn from a larger set of CDR3\eqn{\beta} sequences from
human naive CD8+ T cells.
}
\examples{
data("CDR3ab")
}

\keyword{datasets}
