\name{get_community_feature_purity}
\alias{get_community_feature_purity}
\title{Compute community purity with respect to a node feature}
\description{
Given a \code{node_summary} object returned by \code{detect_communities} with 
column "community", and a node feature, this function computes the purity of 
each community with respect to that feature.

If the node feature is numeric (e.g., expression values, clone size, etc.), 
purity is measured by the coefficient of variation (CV). If the node feature 
is categorical (e.g., tissue type, sample name, cell type), purity is measured 
by Gini Impurity (GI).
}
\details{
The function checks for the existence of a \code{community} column
in \code{node_summary} and for a valid \code{node_feature}. Depending on 
whether the feature is numeric or categorical, it computes CV or GI, 
respectively. 

\itemize{
\item For categorical features: communities with low GI contain 
mostly homogeneous feature values.
\item For numeric features: communities with low CV 
are more homogeneous in feature values.
}
}
\usage{
get_community_feature_purity(node_summary, node_feature)
}
\arguments{
\item{node_summary}{data.frame object generated by \code{detect_communities}.}
\item{node_feature}{Character scalar specifying the vertex attribute to 
evaluate (must be either categorical or numeric).}
}
\value{
A \code{data.frame} summarizing purity for each community:
\itemize{
\item For categorical features: columns \code{GI} (Gini impurity), 
\code{n} (community size), \code{community}.
\item For numeric features: columns \code{mean}, \code{sd}, \code{cv} 
(coefficient of variation), \code{n} (community size), \code{community}.
}
}
\examples{
# load package input data
data("D1", package = "ClustIRR")

# run ClustIRR analysis
c <- clustirr(s = rbind(D1$a, D1$b), meta = rbind(D1$ma, D1$mb))

# detect communities
gcd <- detect_communities(graph = c$graph, 
                          algorithm = "leiden",
                          metric = "average",
                          resolution = 1,
                          weight = "ncweight",
                          iterations = 100,
                          chains = c("CDR3a", "CDR3b"))

# categorical feature
get_community_feature_purity(node_summary = gcd$node_summary, 
                             node_feature = "HLA_A")

# another categorical feature
get_community_feature_purity(node_summary = gcd$node_summary, 
                             node_feature = "TRAV")

# numeric feature
get_community_feature_purity(node_summary = gcd$node_summary, 
                             node_feature = "age")
}
