\name{get_graph}
\alias{get_graph}
\title{Get \code{igraph} object from \code{clust_irr} object}
\usage{get_graph(clust_irr)}
\arguments{
\item{clust_irr}{S4 object generated by the function \code{cluster_irr}}
}
\description{
Given a \code{clust_irr} object generated by the function \code{cluster_irr}, 
the function \code{get_graph} constructs an \code{igraph} object.

The graph nodes represent IR clones. Undirected edges are drawn between 
pairs of nodes, and the attributes of these edges are assigned based 
on the \code{clust_irr} outputs: \eqn{\bar{\omega}}, 
\eqn{\bar{\omega}_c}, etc.
}
\value{
The output is a list with the following elements. First, the list contains 
an \code{igraph} object. The graph nodes and edges contain attributes encoded
in the \code{clust_irr} objects. Second, it contains a data.frame in which 
rows are clones (nodes) in the graph. Third, the list contains the logical 
variable \code{joint_graph}, which is set to \code{TRUE} if the graph is a 
joint graph generated by the function \code{get_joint_graph} and \code{FALSE} 
if the graph is not a joint graph generated by \code{get_graph}.
}
\examples{
# load package input data
data("CDR3ab", package = "ClustIRR")
s <- data.frame(CDR3b = CDR3ab[1:100, "CDR3b"], sample = "A", clone_size = 1)

# run ClustIRR analysis
out <- cluster_irr(s = s)

# get graph
g <- get_graph(clust_irr = out)

names(g)
}
