\name{get_graph}
\alias{get_graph}
\title{Get graph structure from \code{clust_irr} object}
\usage{get_graph(clust_irr)}
\arguments{
\item{clust_irr}{S4 object generated by the function \code{cluster_irr}}
}
\value{
The main output is a list. The list contains an \code{igraph} object. The 
graph vertices and edges contain attributes. Furthermore, it contains a 
data.frame in which rows are clones (vertices) in the graph. Finally, 
the list contains the logical variable \code{joint_graph}, which is set 
to \code{TRUE} if the graph is a joint graph generated by the function 
\code{get_joint_graph} and \code{FALSE} if the graph is not a joint graph 
generated by \code{get_graph}.
}
\description{
As input we take a \code{clust_irr} object generated by the function 
\code{cluster_irr}.

From \code{clust_irr} we generate a graph, in which the nodes are 
clones, and undirected edges are drawn between a pair of vertices 
if the corresponding clones are similar. The main output is an 
\code{igraph} object.
}

\examples{
# load package input data
data("CDR3ab", package = "ClustIRR")
s <- data.frame(CDR3b = CDR3ab[1:100, "CDR3b"], sample = "A", clone_size = 1)

# run ClustIRR analysis
out <- cluster_irr(s = s)

# get graph
g <- get_graph(clust_irr = out)

names(g)
}
