\name{get_graph}
\alias{get_graph}
\title{Get graph structure from \code{clust_irr} object}
\usage{get_graph(clust_irr, sample_id = "S")}
\arguments{
\item{clust_irr}{S4 object generated by the function \code{cluster_irr}}
\item{sample_id}{character, name of the repertoire (default = S)}
}
\value{
The main output of this function is an \code{igraph} object. Additionally, 
we provide a data.frame in which rows are clones (vertices) in the graph.
}

\description{
The main output of this function is an \code{igraph} object.

The vertices in the graph represent clones. Undirected edges are drawn 
between a pair of vertices if the corresponding clones that are locally 
and/or globally similar.
}

\examples{
# load package input data
data("CDR3ab")
s <- base::data.frame(CDR3b = CDR3ab[1:100, "CDR3b"])
r <- base::data.frame(CDR3b = CDR3ab[1:5000, "CDR3b"])

# artificially enrich motif 'RWGW' inside sample dataset
base::substr(x = s$CDR3b[1:20], start = 6, stop = 9) <- "RWGW"

# add an artificial clonal expansion of two sequences to the sample dataset
s <- rbind(s, base::data.frame(CDR3b = rep(x = c("CATSRADKPDGLDALETQYF",
                                                 "CATSRAAKPDGLAALSTQYF"),
                                           times = 5)))

# run ClustIRR analysis
out <- cluster_irr(s = s,
                   r = r,
                   ks = 4,
                   cores = 1,
                   control = list(trim_flank_aa = 3))

# get graph
g <- get_graph(out)

names(g)
}