\name{get_joint_graph}
\alias{get_joint_graph}
\title{Joins two graphs obtained from two \code{clust_irr} objects}
\usage{get_joint_graph(clust_irrs)}
\arguments{
\item{clust_irrs}{A list of at least two S4 objects generated with the 
function \code{cluster_irr}}
}
\value{
The main output of this function is an \code{igraph} object. This object
represents a joint graph of the individual graphs contained as elements 
in the input clust_irrs. One additional output is a data.frame in which 
rows are clones (vertices) in the joint graph.
}

\description{
As input we take at least two \code{clust_irr} objects generated by the 
function \code{cluster_irr}.

Using each \code{clust_irr} object we generate a graph (with the function 
\code{get_graph}) in which the different vertices represent clones, and 
undirected edges are drawn between a pair of vertices if the corresponding 
clones are locally and/or globally similar (see definitions of local/global 
clustering in the documentation of \code{cluster_irr}. 

The function \code{get_joint_graph} performs the following operation for
a pair of graphs:

First, it adds together (union) the vertices. Second, it performs global 
clustering between the two graphs, i.e. it compares the CDR3 sequences of 
the clones between the two graphs. If two clones have similar CDR3 sequences,
then the corresponding vertices are connected by an edge. 

The results is another \code{igraph} object.
}

\examples{
# load package input data
data("CDR3ab")
s <- base::data.frame(CDR3b = CDR3ab[1:100, "CDR3b"])
r <- base::data.frame(CDR3b = CDR3ab[1:5000, "CDR3b"])

# artificially enrich motif 'RWGW' inside sample dataset
base::substr(x = s$CDR3b[1:20], start = 6, stop = 9) <- "RWGW"

# add an artificial clonal expansion of two sequences to the sample dataset
s <- rbind(s, base::data.frame(CDR3b = rep(x = c("CATSRADKPDGLDALETQYF",
                                                 "CATSRAAKPDGLAALSTQYF"),
                                           times = 5)))

# run ClustIRR analysis
c1 <- cluster_irr(s = s, 
                  r = r,
                  version = 2,
                  ks = 4,
                  cores = 1,
                  control = list(trim_flank_aa = 3))
                   
# run ClustIRR analysis
c2 <- cluster_irr(s = s, 
                  r = r,
                  version = 2,
                  ks = 4,
                  cores = 1,
                  control = list(trim_flank_aa = 3))
  
# join clust_irr objects in a list               
clust_irrs <- c(c1, c2)
names(clust_irrs) <- c("C1", "C2")

# get graph
g <- get_joint_graph(clust_irrs = clust_irrs)

names(g)
}