% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs_CD8_ref}
\alias{hs_CD8_ref}
\title{Naive reference repertoire}
\format{
An object of class \code{data.frame} with 573211 rows and 6 columns.
}
\source{
Glanville, Jacob, et al.
"Identifying specificity groups in the T cell receptor repertoire."
Nature 547.7661 (2017): 94.
}
\usage{
data(hs_CD8_ref)
}
\description{
Contains the sequences and V-gene information, V-gene distribution and
CDR3 length distribution.
"human_v2.0_CD8": Reference database of 573,211 CDR3b sequences of
naive human CD48+ T cells provided for GLIPH2.
Additional reference databases were provided for download by the developers
of GLIPH2 in the web tool (http://50.255.35.37:8080).
}
\keyword{datasets}
